pyproject.toml
[project]
name = "openai-to-claude"
version = "1.0.0"
description = "openai to claude"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm[proxy]>=1.80.12",
    "python-dotenv>=1.2.1",
]
